Convert User class to Entity class, add UserRepository class and CustomAuthenticationProvider

+++ src/main/java/com/notes/notesrestservice/User.java
@@ -2,6 +2,11 @@ package com.notes.notesrestservice;
 
 import java.sql.Timestamp;
 
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
 @Entity
 public class User {
     @Id
@@ -16,7 +21,7 @@ public class User {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
